+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- [Added] Graceful handling of exit signals in the default worker
+
 ## v4.0.2 [2023-06-13]
 
 - [Improved] Updated all the dependencies